+++ cub3d/libft/source_files/ft_strlen.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strlen.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: amalbrei <amalbrei@student.42.fr>          +#+  +:+       +#+        */
+/*   By: gchernys <gchernys@42abudhabi.ae>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/01/21 19:45:42 by amalbrei          #+#    #+#             */
-/*   Updated: 2023/03/02 18:09:13 by amalbrei         ###   ########.fr       */
+/*   Updated: 2023/05/15 17:55:48 by gchernys         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,7 +18,7 @@
  * @param c The string to be measured
  * @return size_t The size of the string
  */
-size_t	ft_strlen(const char *c)
+int	ft_strlen(const char *c)
 {
 	size_t	i;
